requirements.txt
torch>=2.0.0
transformers>=4.30.0
gradio>=4.0.0
jinja2>=3.1.0
python-docx>=0.8.11
PyMuPDF>=1.23.0
accelerate>=0.20.0
sentencepiece>=0.1.99
protobuf>=3.20.0